backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
anthropic==0.18.1
python-dotenv==1.0.1
pydantic==2.6.1
starlette==0.36.3
typing-extensions==4.9.0
httpx==0.27.0 